backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
google-genai==0.3.0
Pillow==10.2.0
alembic==1.13.1